feature(refactor-db): detach database from init Log in a single file Test the avro file loading and alert generation from file
