+++ create_records.go
@@ -1,14 +1,9 @@
 package main
 
 import (
-	"context"
 	"io/ioutil"
-	"log"
 	"math"
 	"strconv"
-
-	"go.mongodb.org/mongo-driver/bson"
-	"go.mongodb.org/mongo-driver/mongo/options"
 )
 
 const (
@@ -75,7 +70,6 @@ type AtlasRecord struct {
 var getExtension = map[string]string{
 	"science":    "_istamp.fits.gz",
 	"difference": "_dstamp.fits.gz",
-	"template":   "_rstamp.fits.gz",
 }
 
 func createCutouts(directory string, candid string) map[string]*Cutout {
@@ -204,12 +198,9 @@ func createRecord(data []interface{}) (*AtlasRecord, error) {
 	Publisher := "ATLAS-" + tel
 	Candidate := pCandidate
 	Candid := data[25].(string)
-	ObjectId, err := getOrCreateId(data[25].(string), pCandidate.RA, pCandidate.Dec)
-	if err != nil {
-		return nil, err
-	}
+	ObjectId := data[26].(string)
 
-	// data[26] is mjd,  data[27] is filter, those value goes in the candidate
+	// data[27] is mjd,  data[28] is filter, those value goes in the candidate
 	CutoutScience := data[29].(*Cutout)
 	CutoutDifference := data[30].(*Cutout)
 	CutoutTemplate := data[31].(*Cutout)
@@ -225,109 +216,4 @@ func createRecord(data []interface{}) (*AtlasRecord, error) {
 		CutoutDifference: CutoutDifference,
 	}
 	return atlasRecord, nil
-}
-
-func getOrCreateId(s string, ra float64, dec float64) (string, error) {
-	// get a handle for the trainers collection in the test database
-	collection := client.Database(configuration.Db).Collection(configuration.Col)
-
-	// Find documents
-	// Pass these options to the Find method
-	findOptions := options.Find()
-	findOptions.SetLimit(1)
-	findOptions.SetProjection(bson.D{{Key: "_id", Value: 1}})
-
-	// Passing bson.D{{}} as the filter matches all documents in the collection
-	radius := RADIUS / 3600
-	scaling := wgsScale(dec)
-	meterRadius := radius * scaling
-	lon, lat := ra-180.0, dec
-
-	filter := bson.D{
-		{
-			Key: "loc",
-			Value: bson.D{
-				{
-					Key: "$nearSphere",
-					Value: bson.D{
-						{
-							Key: "$geometry",
-							Value: bson.D{
-								{
-									Key:   "type",
-									Value: "Point",
-								},
-								{
-									Key:   "coordinates",
-									Value: bson.A{lon, lat},
-								},
-							},
-						},
-						{
-							Key:   "$maxDistance",
-							Value: meterRadius,
-						},
-					},
-				},
-			},
-		},
-	}
-
-	cur, err := collection.Find(context.TODO(), filter, findOptions)
-	if err != nil {
-		return s, err
-	}
-
-	// Finding multiple documents returns a cursor
-	// Iterating through the cursor allows us to decode documents one at a time
-	if cur.Next(context.TODO()) {
-
-		// create a value into which the single document can be decoded
-		elem := struct {
-			ObjectId string `bson:"_id"`
-		}{}
-		err := cur.Decode(&elem)
-		if err != nil {
-			return s, err
-		}
-
-		return elem.ObjectId, nil
-	}
-
-	asht := bson.D{
-		{
-			Key: "_id", Value: s,
-		},
-		{
-			Key: "loc", Value: bson.D{
-				{
-					Key: "type", Value: "Point",
-				},
-				{
-					Key: "coordinates", Value: bson.A{lon, lat},
-				},
-			},
-		},
-	}
-
-	_, err = collection.InsertOne(context.TODO(), asht)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	return s, nil
-}
-
-func wgsScale(lat float64) float64 {
-	/*
-		Get scaling to convert degrees to meters at a given geodetic latitude (declination)
-		:param lat: geodetic latitude (declination)
-		:return:
-	*/
-	// Compute radius of curvature along meridian (see https://en.wikipedia.org/wiki/Meridian_arc)
-	rm := a * (1 - math.Pow(e, 2)) / math.Pow(1-math.Pow(e, 2)*math.Pow(math.Sin(lat*math.Pi/180), 2), 1.5)
-
-	// Compute length of arc at this latitude (meters/degree)
-	arc := rm * angle
-	return arc
 }
